refactor: redundant check for invalid action type
Remove an unnecessary require statement for the ActionType passed
in actions to the operate function in the Controller to have the
value of `Call` (7) if no other ActionType enum value is matched.
At runtime, execution is reverted with a Panic code equal to 0x21
("Conversion into non-existent enum type") when an invalid action
type is passed in a call to `operate`, as seen in the accompanying
test case.